testsuite: Add config.libs_have_debug_info
It's not entirely clear that this is the right way to do this since the
compiler we are testing may not have the same configuration as the
current tree. We could add a "Libraries have debugging information"
entry to `ghc --info` but this seems rather heavy just for the
testsuite.